Implemented profile picture upload to S3
Used express-fileupload to get the file blob from the browser, passed
the data property into a buffer, and used that plus uuidv4 to
upload the file to S3 bucket.

Modified the profile.ejs to render the uploaded pic by pulling it from
the Cloudfront distribution.

Added null checking for the rare cases when the browser fails to
upload the image and pass it on.